refactor(agent): add model storing the logic from each json-rpc method and it handles the communication to MongoDB
