+++ modules/agent/historyUtils/getBlockByHash.go
@@ -4,6 +4,7 @@ import (
 	"errors"
 	"eth-agent/common"
 	historyUtilsCommon "eth-agent/modules/agent/historyUtils/common"
+	model "eth-agent/modules/agent/historyUtils/model"
 	"eth-agent/modules/agent/historyUtils/struct/rsps"
 	"fmt"
 
@@ -90,7 +91,7 @@ func GetBlockByHashIndexer(blockHash string, isNeedAllTx bool) interface{} {
 		"hash": blockHash,
 	}
 
-	result, err := RetrieveBlock(condition)
+	result, err := model.RetrieveBlock(condition)
 	if err != nil {
 		errors := common.Error{
 			ErrorType:        1,
@@ -172,33 +173,3 @@ func GetBlockByHashIndexer(blockHash string, isNeedAllTx bool) interface{} {
 	responseBlockTxHashOnlyType.Result.Transactions = txsHash
 	return responseBlockTxHashOnlyType
 }
-
-// redeclare in getBlockTransactionCountByNumber
-
-// RetrieveBlock retrieve specific block data from mongo
-// func RetrieveBlock(conditions map[string]interface{}) ([]blockStrcut.Block, error) {
-// 	var err error
-
-// 	mongo, err := historyUtilsMongo.GetMongoSession()
-// 	if err != nil {
-// 		errors := common.Error{
-// 			ErrorType:        1,
-// 			ErrorDescription: err.Error(),
-// 		}
-// 		logger.Console().Panic(errors)
-// 		logger.File().Error(err)
-// 	}
-
-// 	defer mongo.Close()
-
-// 	collection := mongo.DB(dbName).C(BS_Blocks)
-// 	result := []blockStrcut.Block{}
-// 	err = collection.Find(conditions).All(&result)
-
-// 	if err != nil {
-// 		message := fmt.Sprintf("Retrive Block failded")
-// 		err = errors.New(message)
-// 	}
-
-// 	return result, err
-// }
