+++ modules/agent/historyUtils/getTxReceipt.go
@@ -3,10 +3,8 @@ package historyUtils
 import (
 	"errors"
 	"eth-agent/common"
-	collectionName "eth-agent/modules/agent/historyUtils/common"
 	historyUtilsCommon "eth-agent/modules/agent/historyUtils/common"
-	historyUtilsMongo "eth-agent/modules/agent/historyUtils/mongo"
-	receiptStruct "eth-agent/modules/agent/historyUtils/struct/bs_receipt"
+	model "eth-agent/modules/agent/historyUtils/model"
 	"eth-agent/modules/agent/historyUtils/struct/rsps"
 	"fmt"
 
@@ -59,7 +57,7 @@ func GetTxReceiptIndexer(transactionAddress string) rsps.ReceiptReponse {
 		"transactionHash": transactionAddress,
 	}
 
-	result, err := RetrieveReceipt(condition)
+	result, err := model.RetrieveReceipts(condition)
 	if err != nil {
 		errors := common.Error{
 			ErrorType:        1,
@@ -75,32 +73,4 @@ func GetTxReceiptIndexer(transactionAddress string) rsps.ReceiptReponse {
 
 	response.Result = result[0]
 	return response
-}
-
-// RetrieveBlock retrieve specific block data from mongo
-func RetrieveReceipt(conditions map[string]interface{}) ([]receiptStruct.Receipt, error) {
-	var err error
-
-	mongo, err := historyUtilsMongo.GetMongoSession()
-	if err != nil {
-		errors := common.Error{
-			ErrorType:        1,
-			ErrorDescription: err.Error(),
-		}
-		logger.Console().Panic(errors)
-		logger.File().Error(err)
-	}
-
-	defer mongo.Close()
-
-	collection := mongo.DB(dbName).C(collectionName.BsReceipts)
-	result := []receiptStruct.Receipt{}
-	err = collection.Find(conditions).All(&result)
-
-	if err != nil {
-		message := fmt.Sprintf("Retrive Receipt of Transaction failded")
-		err = errors.New(message)
-	}
-
-	return result, err
 }
